+++ out/extension.js
@@ -3,6 +3,7 @@ Object.defineProperty(exports, "__esModule", { value: true });
 exports.activate = activate;
 const crypto_1 = require("crypto");
 const vscode = require("vscode");
+const alertMessage_1 = require("./alertMessage");
 function activate(context) {
     const provider = new SidebarProvider(context.extensionUri);
     context.subscriptions.push(vscode.window.registerWebviewViewProvider('sidebarWebviewView', provider));
@@ -18,7 +19,7 @@ class SidebarProvider {
         };
         webviewView.webview.html = this._getHtmlForWebview(webviewView.webview);
         webviewView.webview.onDidReceiveMessage(message => {
-            const alert = parseAlertMessage(message);
+            const alert = (0, alertMessage_1.parseAlertMessage)(message);
             if (alert) {
                 vscode.window.showInformationMessage(alert.text);
             }
@@ -50,28 +51,6 @@ class SidebarProvider {
 </body>
 </html>`;
     }
-}
-function parseAlertMessage(message) {
-    if (!message || typeof message !== 'object' || Array.isArray(message)) {
-        return undefined;
-    }
-    const prototype = Object.getPrototypeOf(message);
-    if (prototype !== Object.prototype && prototype !== null) {
-        return undefined;
-    }
-    const candidate = message;
-    if (!Object.prototype.hasOwnProperty.call(candidate, 'command') ||
-        !Object.prototype.hasOwnProperty.call(candidate, 'text')) {
-        return undefined;
-    }
-    if (candidate.command !== 'alert' || typeof candidate.text !== 'string') {
-        return undefined;
-    }
-    const text = candidate.text.trim();
-    if (text.length === 0 || text.length > 200 || /[\r\n]/.test(text)) {
-        return undefined;
-    }
-    return { command: 'alert', text };
 }
 function getNonce() {
     return (0, crypto_1.randomBytes)(16).toString('base64');